Enable builds on windows
