+++ Makefile
@@ -12,6 +12,8 @@ build.lib:
 	@GOOS=linux GOARCH=amd64 go build ./...
 	@GOOS=darwin GOARCH=arm64 go build ./...
 	@GOOS=darwin GOARCH=amd64 go build ./...
+	@GOOS=windows GOARCH=arm64 go build ./...
+	@GOOS=windows GOARCH=amd64 go build ./...
 
 .PHONY: build.bench
 build.bench:
