+++ wasm/jit/mmap_linux.go
@@ -1,5 +1,24 @@
+//go:build linux
+// +build linux
+
 package jit
 
 import "syscall"
 
 const mmapFlags = syscall.MAP_ANONYMOUS
+
+// Copy the code into the executable region
+// and returns the byte slice of the region.
+func mmapCodeSegment(code []byte) ([]byte, error) {
+	mmapFunc, err := syscall.Mmap(
+		-1,
+		0,
+		len(code),
+		syscall.PROT_READ|syscall.PROT_WRITE|syscall.PROT_EXEC, syscall.MAP_PRIVATE|mmapFlags,
+	)
+	if err != nil {
+		return nil, err
+	}
+	copy(mmapFunc, code)
+	return mmapFunc, nil
+}
